fixed warning about uint copared to int; added .tar.gz for install.package() installation;

+++ src/multbxxc.cpp
@@ -427,7 +427,7 @@ void resize(SEXP& x_, uvec& di) {
    }
 #endif
    //Rcout << "n=" << XLENGTH(x_) << endl;
-   unsigned int pdi=prod(di);
+   int pdi= (int) prod(di);
    if (TRUELENGTH(x_) >= pdi)
       SETLENGTH(x_, pdi);
    else
